mcp auth wip
