+++ libs/model/src/aggregates/mcp/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateKlavisMCPInstance.command';
