+++ libs/schemas/src/commands/index.ts
@@ -6,6 +6,7 @@ export * from './comment.schemas';
 export * from './community.schemas';
 export * from './contest.schemas';
 export * from './discord.schemas';
+export * from './integrations.schemas';
 export * from './load-testing.schemas';
 export * from './magic.schemas';
 export * from './poll.schemas';
