+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Integrations/GoogleSheetsMCP/index.tsx
@@ -1,0 +1,3 @@
+import GoogleSheetsMCP from './GoogleSheetsMCP';
+
+export default GoogleSheetsMCP;
