+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Integrations/Integrations.tsx
@@ -7,6 +7,7 @@ import CustomTOS from './CustomTOS';
 import CustomURL from './CustomURL';
 import Directory from './Directory';
 import Discord from './Discord';
+import GoogleSheetsMCP from './GoogleSheetsMCP';
 import './Integrations.scss';
 import OnchainVerification from './OnchainVerification';
 import Snapshots from './Snapshots';
@@ -28,7 +29,7 @@ const Integrations = () => {
       featureHint={{
         title: 'Everything in one place',
         description: `
-         You can link to your projects custom terms of service page. 
+         You can link to your projects custom terms of service page.
          You can also contact us to create a custom URL that points to your Common community.
             `,
       }}
@@ -42,6 +43,7 @@ const Integrations = () => {
         {isEthereum && <Snapshots />}
         <Tokenization />
         <Discord />
+        <GoogleSheetsMCP />
         <Webhooks />
         <CustomTOS />
         <CustomURL />
