+++ packages/commonwealth/server/api/index.ts
@@ -2,3 +2,4 @@ export * as external from './external-router';
 export * as integration from './integration-router';
 export * as internal from './internal-router';
 export * as mcp from './mcp';
+export * as mcpServer from './mcp-server';
